[DSE] Translate killing locations through phis.
This patch updates DSE to translate the address of the killing access
through phis when traversing MemoryPhis. Note that this still misses
cases where we could translate phis, but there are no MemoryPhis in the
block with the translatable phi.

This is an initial step to improve DSE in the presence of address phis.
It can be improved in the future to also do phi translation when
stepping over any type of access, not just MemoryPhis. There's also a
related patch D90095.

Differential Revision: https://reviews.llvm.org/D98288