[IndVars] Use exit count up to current block when handing ANDs/ORs
When analyzing exit conditions joined through AND/OR, use exit count
computed not for current block, but for all blocks from loop start and up
to the current block together. It is not actually required to know that the
current block will exit to skip the last iteration optimistically, it is enough
to know that we will exit on this block or earlier.

Differential Revision: https://reviews.llvm.org/D141594