Clang CodeGen for restrict-qualified pointers (Based on D9403)
Introduce support for restrict-qualified pointers as local variables,
as member of a struct, arrays of restrict-qualified pointers,
pointers to restrict qualified pointers.

The full support is enabled with '-full-restrict'.
Combine it with '-no-noalias-arguments' to avoid putting the 'noalias'
attribute on restrict pointers used as argument. This combiniation is
needed to correctly support restrict arguments.

[noalias] D9403: llvm.noalias - Clang CodeGen for local restrict-qualified pointers

Rebase of https://reviews.llvm.org/D9403

[clang] switch to NoAliasOffsets

[restrict] FIXUP Also support base classes with restrict