[VPlan] Move mask handling to VPRecipeBase (NFCI).
Consolidate mask handling to VPRecipeBase, buy letting subclasses define
the number of non-mask operations, if it is different than the total
number of operands.

This removes a bit of duplication, allows for generic printing of the
mask and potentially allows masking of any recipe (e.g. to use
vector predication).

Differential Revision: https://reviews.llvm.org/D157442