[VPlan] Build initial VPlan 0 using HCFGBuilder for inner loops. (NFC)
Use HCFGBuilder to build an initial VPlan 0, which wraps all input
instructions in VPInstructions and update tryToBuildVPlanWithVPRecipes
to replace the VPInstructions with widened recipes.

At the moment, widened recipes are created based on the underlying
instruction of the VPInstruction. Masks are also still created based on
the input IR basic blocks and the loop CFG is flattened in the main
loop processing the VPInstructions.

This patch also incldues support for Switch instructions in HCFGBuilder
using just a VPInstruction with Instruction::Switch opcode.

There are multiple follow-ups planned:
 * Use VPIRInstructions instead of VPInstructions in HCFGBuilder,
 * Perform predication on the VPlan directly,
 * Unify code constructing VPlan 0 to be shared by both inner and outer
   loop code paths.